requirements.txt
astroid==2.1.0
certifi==2018.11.29
chardet==3.0.4
coverage==4.5.2
coveralls==1.5.1
Django==2.1.4
django-nose==1.4.6
djangorestframework==3.9.0
djangorestframework-camel-case==0.2.0
djangorestframework-jwt==1.11.0
docopt==0.6.2
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
nose==1.3.7
psycopg2==2.7.6.1
PyJWT==1.7.0
pylint==2.2.2
pytz==2018.7
requests==2.21.0
six==1.12.0
typed-ast==1.1.1
urllib3==1.24.1
wrapt==1.10.11
